Moved deserialization code to serializer.
