+++ AtomEventStore/ConventionBasedSerializerOfComplexImmutableClasses.cs
@@ -1,10 +1,13 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 using System.Xml;
+using System.Xml.Linq;
 
 namespace Grean.AtomEventStore
 {
@@ -100,6 +103,165 @@ namespace Grean.AtomEventStore
                 && type != typeof(int)
                 && type != typeof(long)
                 && type != typeof(string);
+        }
+
+        public XmlAtomContent Deserialize(XmlReader xmlReader)
+        {
+            if (xmlReader == null)
+                throw new ArgumentNullException("xmlReader");
+
+            xmlReader.MoveToContent();
+            var x = (XElement)XElement.ReadFrom(xmlReader);
+            var item = ReadFrom(x.Elements().Single());
+            return new XmlAtomContent(item);
+        }
+
+        private static object ReadFrom(XElement node)
+        {
+            if (!node.HasElements)
+                return node.Value;
+
+            var elementName = node.Name.LocalName;
+            var xmlNamespace = node.Name.NamespaceName;
+
+            var dotNetNamespace = UnUrnify(xmlNamespace);
+            var itemType = new XmlCasedName(elementName).ToTypeIn(dotNetNamespace);
+
+            var ctor = GetMostModestConstructor(itemType);
+
+            var namedArguments = (from p in ctor.GetParameters()
+                                  let xpn = ConventionBasedSerializerOfComplexImmutableClasses.Xmlify(p.Name)
+                                  join x in node.Elements() on xpn equals x.Name.LocalName
+                                  select GetObjectFrom(x))
+                                 .ToList();
+            var paramsValues = (from x in node.Elements()
+                                where !namedArguments.Select(a => a.Name).Contains(x.Name.LocalName)
+                                select GetParamsObjectFrom(x))
+                               .ToArray();
+
+            if (itemType.IsGenericTypeDefinition)
+            {
+                var matchingArgumentType = GetMatchingArgument(namedArguments, paramsValues, itemType, ctor);
+
+                itemType = itemType.MakeGenericType(matchingArgumentType);
+                ctor = GetMostModestConstructor(itemType);
+            }
+
+            List<object> sortedArguments = new List<object>();
+            foreach (var p in ctor.GetParameters())
+            {
+                if (Attribute.IsDefined(p, typeof(ParamArrayAttribute)))
+                {
+                    var elementType = p.ParameterType.GetElementType();
+                    var arr = Array.CreateInstance(elementType, paramsValues.Length);
+                    paramsValues.Select(a => a.Value).ToArray().CopyTo(arr, 0);
+                    sortedArguments.Add(arr);
+                }
+                else
+                {
+                    var argValue = namedArguments.Single(kvp => UnXmlify(kvp.Name).Equals(p.Name, StringComparison.OrdinalIgnoreCase));
+                    sortedArguments.Add(ChangeType(argValue.Value, p.ParameterType));
+                }
+            }
+
+            var item = ctor.Invoke(sortedArguments.ToArray());
+            return item;
+        }
+
+        private static string UnUrnify(string text)
+        {
+            return string.Join(".", text.Replace("urn:", "").Split(':').Select(UnXmlify));
+        }
+
+        private static string UnXmlify(string text)
+        {
+            return new XmlCasedName(text).ToPascalCase();
+        }
+
+        private static ConstructorInfo GetMostModestConstructor(Type itemType)
+        {
+            return (from c in itemType.GetConstructors()
+                    let args = c.GetParameters()
+                    orderby args.Length
+                    select c).First();
+        }
+
+        private class Argument
+        {
+            public readonly string Name;
+            public readonly object Value;
+
+            public Argument(string name, object value)
+            {
+                this.Name = name;
+                this.Value = value;
+            }
+        }
+
+        private static Argument GetParamsObjectFrom(XElement node)
+        {
+            var value = ReadFrom(node);
+            return new Argument(node.Name.LocalName, value);
+        }
+
+        private static Argument GetObjectFrom(XElement node)
+        {
+            if (!node.HasElements)
+                return new Argument(node.Name.LocalName, node.Value);
+
+            var value = ReadFrom(node.Elements().Single());
+            return new Argument(node.Name.LocalName, value);
+        }
+
+        private static Type GetMatchingArgument(List<Argument> namedArguments, Argument[] paramsValues, Type itemType, ConstructorInfo ctor)
+        {
+            return GetDirectMatchingArgument(namedArguments, itemType, ctor).Concat(
+                GetArrayMatchingArgument(paramsValues))
+                .First();
+        }
+
+        private static IEnumerable<Type> GetDirectMatchingArgument(List<Argument> arguments, Type itemType, ConstructorInfo ctor)
+        {
+            var openGenericType = itemType.GetGenericArguments().Single();
+            var genericCtorParamter = ctor.GetParameters()
+                .SingleOrDefault(p => p.ParameterType == openGenericType);
+            if (genericCtorParamter == null)
+                yield break;
+
+            var matchingArgument = arguments
+                .Single(kvp => UnXmlify(kvp.Name).Equals(genericCtorParamter.Name, StringComparison.OrdinalIgnoreCase))
+                .Value;
+            yield return matchingArgument.GetType();
+        }
+
+        private static IEnumerable<Type> GetArrayMatchingArgument(IEnumerable<Argument> paramsValues)
+        {
+            var interfaces = from a in paramsValues
+                             select GetBaseTypes(a.Value.GetType());
+
+            var commonInterfaces = interfaces.Aggregate((x, y) => x.Intersect(y).ToArray());
+            var commonInterface = commonInterfaces.FirstOrDefault();
+            if (commonInterface == null)
+                yield break;
+
+            yield return commonInterface;
+        }
+
+        private static IEnumerable<Type> GetBaseTypes(Type type)
+        {
+            return new[] { type }.Concat(type.GetInterfaces());
+        }
+
+        private static object ChangeType(object value, Type type)
+        {
+            if (type == typeof(Guid))
+                return (Guid)UuidIri.Parse(value.ToString());
+            if (type == typeof(Uri))
+                return new Uri(value.ToString());
+            if (type == typeof(DateTimeOffset))
+                return DateTimeOffset.Parse(value.ToString(), CultureInfo.InvariantCulture);
+
+            return Convert.ChangeType(value, type, CultureInfo.InvariantCulture);
         }
     }
 }
